Refactor: Update Next.js, dependencies, and page structure
This commit updates Next.js to version 14, refactors the project structure to use the `app` directory, and updates various dependencies. It also removes old `pages` directory files and adjusts imports accordingly.


[2025-09-22 20:23:53]
