+++ app/api/health/route.ts
@@ -1,0 +1,11 @@
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  const response = NextResponse.json(
+    { status: 'ok', timestamp: new Date().toISOString() },
+    { status: 200 }
+  );
+  
+  response.headers.set('Cache-Control', 'no-store');
+  return response;
+}
