+++ app/api/status/route.ts
@@ -1,0 +1,8 @@
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  return NextResponse.json(
+    { status: 'ok', timestamp: new Date().toISOString() },
+    { status: 200 }
+  );
+}
