+++ app/blog/ai-2026-multimodal-agents-enterprise-blueprint/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Link from 'next/link';
-import SEO from '../../../components/SEO';
+import SEO from '../../components/SEO';
 
 export const metadata = {
   title: 'AI 2026: Multimodal Agents Enterprise Blueprint | Zion Tech Group',
