+++ app/content/ai-2026-evaluation-hub/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Link from 'next/link';
-import SEO from '../../../components/SEO';
+import SEO from '../../components/SEO';
 
 export const metadata = {
   title: 'AI 2026: Evaluation Hub | Zion Tech Group',
