+++ app/content/ai-2026-executive-guide-to-agent-adoption/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Link from 'next/link';
-import SEO from '../../../components/SEO';
+import SEO from '../../components/SEO';
 
 export const metadata = {
   title: 'AI 2026: Executive Guide to Agent Adoption | Zion Tech Group',
