+++ app/content/ai-2026-go-to-market-agents-blueprint/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SEO from '../../../components/SEO';
+import SEO from '../../components/SEO';
 import Link from 'next/link';
 
 export const metadata = {
