+++ app/content/edge-ai-agents-2026/page.tsx
@@ -1,7 +1,6 @@
 import Link from 'next/link';
 import React from 'react';
-import { ErrorBoundary } from '../../../components/ErrorBoundary';
-import SEO from '../../components/SEO';
+import ErrorBoundary from '../../components/ErrorBoundary';
 
 export const metadata = {
   title: 'Edge AI Agents 2026 - Real-Time Autonomy Guide',
@@ -13,12 +12,6 @@ export default function EdgeAIAgents2026() {
   return (
     <ErrorBoundary>
       <div className="min-h-screen bg-white">
-        <SEO
-          title="Edge AI Agents 2026: Real-Time Autonomy"
-          description="How edge-native AI agents enable ultra-low-latency decisioning and control with on-device intelligence in 2026."
-          keywords="edge AI, AI agents, real-time autonomy, on-device AI"
-          url="/content/edge-ai-agents-2026"
-        />
 
         <main className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 py-16">
           <h1 className="text-4xl font-bold text-gray-900 mb-6">Edge AI Agents 2026: Real-Time Autonomy</h1>
