+++ netlify.toml
@@ -7,6 +7,7 @@
 [build.environment]
   NODE_VERSION = "20"
   NPM_FLAGS = "--legacy-peer-deps"
+  NODE_OPTIONS = "--max-old-space-size=4096"
 
 # Use the official Netlify Next.js plugin for SSR/ISR support
 [[plugins]]
