+++ package.json
@@ -8,14 +8,14 @@
   },
   "scripts": {
     "dev": "next dev",
-    "build": "NODE_OPTIONS=--openssl-legacy-provider next build --no-lint",
+    "build": "NODE_OPTIONS='--max-old-space-size=4096 --openssl-legacy-provider' next build --no-lint",
     "start": "next start",
     "lint": "next lint",
     "build:analyze": "ANALYZE=true npm run build",
     "build:clean": "rm -rf .next && npm run build",
     "dev:clean": "rm -rf .next && npm run dev",
     "lint:fix": "next lint --fix",
-    "type-check": "tsc --noEmit",
+    "type-check": "NODE_OPTIONS='--max-old-space-size=4096' tsc --noEmit",
     "test": "jest",
     "test:watch": "jest --watch",
     "preview": "npm run build && npm run start",
@@ -34,20 +34,20 @@
     "clsx": "2.1.1",
     "framer-motion": "^12.23.12",
     "lucide-react": "^0.536.0",
-    "next": "12.3.4",
+    "next": "^14.2.0",
     "next-themes": "^0.4.6",
     "node-cron": "^4.2.1",
     "openai": "^5.11.0",
     "path": "^0.12.7",
-    "react": "^17.0.2",
+    "react": "^18.3.0",
     "react-datepicker": "^8.4.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.0",
     "sonner": "^2.0.7",
     "swr": "^2.3.6",
     "tailwind-merge": "^3.3.1",
     "uuid": "^11.1.0",
-    "@types/react": "^17.0.88",
-    "@types/react-dom": "^17.0.26"
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0"
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^14.2.0",
