--- pages/_document.jsx
@@ -1,22 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
-
-class MyDocument extends Document {
-  render() {
-    return (
-      <Html lang="en">
-        <Head>
-          <meta name="description" content="Zion Tech Group — AI, Cloud, and Security services." />
-          <meta name="theme-color" content="#0ea5e9" />
-          <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="" />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
-
