--- pages/api/health.js
@@ -1,4 +1,0 @@
-export default function handler(req, res) {
-  res.setHeader('Cache-Control', 'no-store');
-  res.status(200).json({ status: 'ok', timestamp: new Date().toISOString() });
-}
