--- pages/contact.jsx
@@ -1,305 +1,0 @@
-import React, { useState } from 'react'
-import Head from 'next/head'
-import Link from 'next/link'
-
-function Contact() {
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    company: '',
-    message: '',
-    service: ''
-  })
-  const [formErrors, setFormErrors] = useState({})
-  const [isSubmitting, setIsSubmitting] = useState(false)
-  const [submitSuccess, setSubmitSuccess] = useState(false)
-
-  const validateForm = () => {
-    const errors = {}
-    
-    if (!formData.name.trim()) {
-      errors.name = 'Name is required'
-    }
-    
-    if (!formData.email.trim()) {
-      errors.email = 'Email is required'
-    } else if (!/\S+@\S+\.\S+/.test(formData.email)) {
-      errors.email = 'Please enter a valid email address'
-    }
-    
-    if (!formData.message.trim()) {
-      errors.message = 'Message is required'
-    }
-    
-    setFormErrors(errors)
-    return Object.keys(errors).length === 0
-  }
-
-  const handleChange = (e) => {
-    const { name, value } = e.target
-    setFormData({
-      ...formData,
-      [name]: value
-    })
-    
-    // Clear error when user starts typing
-    if (formErrors[name]) {
-      setFormErrors({
-        ...formErrors,
-        [name]: ''
-      })
-    }
-  }
-
-  const handleSubmit = async (e) => {
-    e.preventDefault()
-    
-    if (!validateForm()) {
-      return
-    }
-    
-    setIsSubmitting(true)
-    
-    try {
-      // Simulate API call
-      await new Promise(resolve => setTimeout(resolve, 1000))
-      
-      console.log('Form submitted:', formData)
-      setSubmitSuccess(true)
-      setFormData({ name: '', email: '', company: '', message: '', service: '' })
-      setFormErrors({})
-      
-      // Reset success message after 5 seconds
-      setTimeout(() => setSubmitSuccess(false), 5000)
-    } catch (error) {
-      console.error('Error submitting form:', error)
-    } finally {
-      setIsSubmitting(false)
-    }
-  }
-
-  return (
-    <div className="min-h-screen bg-gradient-to-br from-slate-900 via-purple-900 to-slate-900">
-      <Head>
-        <title>Contact Us — Zion Tech Group</title>
-        <meta name="description" content="Get in touch with Zion Tech Group for your IT and AI service needs. We're here to help transform your business." />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-      </Head>
-
-      {/* Navigation */}
-      <nav className="fixed top-0 w-full z-50 bg-black/20 backdrop-blur-md border-b border-white/10">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center py-4">
-            <Link href="/" className="text-2xl font-bold text-white">Zion Tech Group</Link>
-            <div className="hidden md:flex space-x-8">
-              <Link href="/about" className="text-white hover:text-purple-300 transition-colors">About</Link>
-              <Link href="/services" className="text-white hover:text-purple-300 transition-colors">Services</Link>
-              <Link href="/blog" className="text-white hover:text-purple-300 transition-colors">Blog</Link>
-              <Link href="/contact" className="text-purple-300 font-semibold">Contact</Link>
-            </div>
-          </div>
-        </div>
-      </nav>
-
-      <main className="pt-20">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-20">
-          {/* Hero Section */}
-          <div className="text-center mb-20">
-            <h1 className="text-5xl md:text-6xl font-bold text-white mb-6">
-              Contact <span className="bg-gradient-to-r from-purple-400 to-pink-400 bg-clip-text text-transparent">Us</span>
-            </h1>
-            <p className="text-xl text-gray-300 max-w-3xl mx-auto">
-              Ready to transform your business with cutting-edge technology? Let's discuss your project and how we can help you achieve your goals.
-            </p>
-          </div>
-
-          <div className="grid lg:grid-cols-2 gap-12">
-            {/* Contact Form */}
-            <div className="bg-white/10 backdrop-blur-md rounded-2xl p-8 border border-white/20">
-              <h2 className="text-3xl font-bold text-white mb-6">Send us a Message</h2>
-              <form onSubmit={handleSubmit} className="space-y-6">
-                <div className="grid md:grid-cols-2 gap-6">
-                  <div>
-                    <label htmlFor="name" className="block text-white font-semibold mb-2">Name *</label>
-                    <input
-                      type="text"
-                      id="name"
-                      name="name"
-                      value={formData.name}
-                      onChange={handleChange}
-                      required
-                      className={`w-full px-4 py-3 bg-white/10 border rounded-lg text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:border-transparent ${
-                        formErrors.name ? 'border-red-500 focus:ring-red-500' : 'border-white/20 focus:ring-purple-500'
-                      }`}
-                      placeholder="Your full name"
-                    />
-                    {formErrors.name && <p className="text-red-400 text-sm mt-1">{formErrors.name}</p>}
-                  </div>
-                  <div>
-                    <label htmlFor="email" className="block text-white font-semibold mb-2">Email *</label>
-                    <input
-                      type="email"
-                      id="email"
-                      name="email"
-                      value={formData.email}
-                      onChange={handleChange}
-                      required
-                      className={`w-full px-4 py-3 bg-white/10 border rounded-lg text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:border-transparent ${
-                        formErrors.email ? 'border-red-500 focus:ring-red-500' : 'border-white/20 focus:ring-purple-500'
-                      }`}
-                      placeholder="your.email@company.com"
-                    />
-                    {formErrors.email && <p className="text-red-400 text-sm mt-1">{formErrors.email}</p>}
-                  </div>
-                </div>
-                
-                <div>
-                  <label htmlFor="company" className="block text-white font-semibold mb-2">Company</label>
-                  <input
-                    type="text"
-                    id="company"
-                    name="company"
-                    value={formData.company}
-                    onChange={handleChange}
-                    className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-lg text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-purple-500 focus:border-transparent"
-                    placeholder="Your company name"
-                  />
-                </div>
-
-                <div>
-                  <label htmlFor="service" className="block text-white font-semibold mb-2">Service Interest</label>
-                  <select
-                    id="service"
-                    name="service"
-                    value={formData.service}
-                    onChange={handleChange}
-                    className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-lg text-white focus:outline-none focus:ring-2 focus:ring-purple-500 focus:border-transparent"
-                  >
-                    <option value="">Select a service</option>
-                    <option value="ai-ml">AI & Machine Learning</option>
-                    <option value="cloud">Cloud Infrastructure</option>
-                    <option value="cybersecurity">Cybersecurity</option>
-                    <option value="devops">DevOps & Automation</option>
-                    <option value="consulting">Technology Consulting</option>
-                    <option value="other">Other</option>
-                  </select>
-                </div>
-
-                <div>
-                  <label htmlFor="message" className="block text-white font-semibold mb-2">Message *</label>
-                  <textarea
-                    id="message"
-                    name="message"
-                    value={formData.message}
-                    onChange={handleChange}
-                    required
-                    rows={5}
-                    className={`w-full px-4 py-3 bg-white/10 border rounded-lg text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:border-transparent ${
-                      formErrors.message ? 'border-red-500 focus:ring-red-500' : 'border-white/20 focus:ring-purple-500'
-                    }`}
-                    placeholder="Tell us about your project and how we can help..."
-                  />
-                  {formErrors.message && <p className="text-red-400 text-sm mt-1">{formErrors.message}</p>}
-                </div>
-
-                {submitSuccess && (
-                  <div className="bg-green-600/20 border border-green-500/50 text-green-400 px-4 py-3 rounded-lg">
-                    Thank you for your message! We'll get back to you soon.
-                  </div>
-                )}
-                
-                <button
-                  type="submit"
-                  disabled={isSubmitting}
-                  className={`w-full py-4 rounded-lg font-semibold transition-all transform ${
-                    isSubmitting
-                      ? 'bg-gray-600 text-gray-300 cursor-not-allowed'
-                      : 'bg-gradient-to-r from-purple-600 to-pink-600 text-white hover:from-purple-700 hover:to-pink-700 hover:scale-105'
-                  }`}
-                >
-                  {isSubmitting ? 'Sending...' : 'Send Message'}
-                </button>
-              </form>
-            </div>
-
-            {/* Contact Information */}
-            <div className="space-y-8">
-              <div className="bg-white/10 backdrop-blur-md rounded-2xl p-8 border border-white/20">
-                <h3 className="text-2xl font-bold text-white mb-6">Get in Touch</h3>
-                <div className="space-y-4">
-                  <div className="flex items-center">
-                    <div className="text-purple-400 mr-4">📧</div>
-                    <div>
-                      <div className="text-white font-semibold">Email</div>
-                      <div className="text-gray-300">contact@ziontechgroup.com</div>
-                    </div>
-                  </div>
-                  <div className="flex items-center">
-                    <div className="text-purple-400 mr-4">📞</div>
-                    <div>
-                      <div className="text-white font-semibold">Phone</div>
-                      <div className="text-gray-300">+1 (555) 123-4567</div>
-                    </div>
-                  </div>
-                  <div className="flex items-center">
-                    <div className="text-purple-400 mr-4">📍</div>
-                    <div>
-                      <div className="text-white font-semibold">Office</div>
-                      <div className="text-gray-300">San Francisco, CA</div>
-                    </div>
-                  </div>
-                  <div className="flex items-center">
-                    <div className="text-purple-400 mr-4">⏰</div>
-                    <div>
-                      <div className="text-white font-semibold">Business Hours</div>
-                      <div className="text-gray-300">Mon - Fri: 9:00 AM - 6:00 PM PST</div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div className="bg-gradient-to-r from-purple-600/20 to-pink-600/20 backdrop-blur-md rounded-2xl p-8 border border-white/20">
-                <h3 className="text-2xl font-bold text-white mb-4">Why Choose Us?</h3>
-                <ul className="space-y-3 text-gray-300">
-                  <li className="flex items-center">
-                    <span className="text-purple-400 mr-3">✓</span>
-                    24/7 Support & Monitoring
-                  </li>
-                  <li className="flex items-center">
-                    <span className="text-purple-400 mr-3">✓</span>
-                    Proven Track Record
-                  </li>
-                  <li className="flex items-center">
-                    <span className="text-purple-400 mr-3">✓</span>
-                    Custom Solutions
-                  </li>
-                  <li className="flex items-center">
-                    <span className="text-purple-400 mr-3">✓</span>
-                    Competitive Pricing
-                  </li>
-                </ul>
-              </div>
-            </div>
-          </div>
-        </div>
-      </main>
-
-      {/* Footer */}
-      <footer className="bg-black/40 backdrop-blur-md border-t border-white/10">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-          <div className="text-center text-gray-400">
-            <p>&copy; 2025 Zion Tech Group. All rights reserved.</p>
-            <div className="mt-4 flex justify-center space-x-6">
-              <Link href="/about" className="hover:text-white transition-colors">About</Link>
-              <Link href="/services" className="hover:text-white transition-colors">Services</Link>
-              <Link href="/blog" className="hover:text-white transition-colors">Blog</Link>
-              <Link href="/contact" className="hover:text-white transition-colors">Contact</Link>
-            </div>
-          </div>
-        </div>
-      </footer>
-    </div>
-  )
-}
-
-export default Contact
